sh: Move fragments of shell settings to .config/sh/shrc.d
.profile is not read when opening wayland session.  So dynamically set
environment variables by .profile.d/*.sh is not set when launching a new
terminal on wayland.

systemd.environment-generator can be used in this situation, which is
dynamically generate environment settings, but it cannot be used as user
config.

The surest way to set dynamic environment variables is loading the
fragments from .zshrc.